requirements.txt
absl-py==2.3.1
appdirs==1.4.4
asgiref==3.9.0
astunparse==1.6.3
black==25.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
fonttools==4.58.5
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
protobuf==5.29.5
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.1.96
